+++ doc/acrn.doxyfile
@@ -795,10 +795,10 @@ WARN_LOGFILE           =
 INPUT                  =  custom-doxygen/mainpage.md \
                           ../hypervisor/include/arch/x86/guest/guest.h \
                           ../hypervisor/include/arch/x86/mmu.h \
-                          ../hypervisor/include/arch/x86/mtrr.h \
                           ../hypervisor/include/arch/x86/pgtable.h \
                           ../hypervisor/include/arch/x86/vtd.h \
                           ../hypervisor/include/arch/x86/irq.h \
+                          ../hypervisor/include/arch/x86/guest/vmtrr.h \
                           ../hypervisor/include/arch/x86/guest/vlapic.h \
                           ../hypervisor/include/arch/x86/guest/vioapic.h \
                           ../hypervisor/include/arch/x86/guest/vpic.h \
